requirements.txt
aiogram
python-dotenv
gspread
oauth2client